% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_DB_pulls.R
\name{dbPull}
\alias{dbPull}
\title{Pull data from relational data base.}
\usage{
dbPull(vSelect = NULL, filePath)
}
\arguments{
\item{vSelect}{Character vector of variables that should be pulled from data base.}

\item{filePath}{Path of the existing db file.}
}
\value{
Returns a data frame, including the relevant variables.
}
\description{
Function to extract specific (or all) variables from various data tables. Variables are merged in the specified merge order via left joins and using the foreign keys.
}
\details{
Note that the exact merging process is determined when the data base is created via createDB and can not be altered posthoc. For alternative forms of full requestes please contact the package author as the package is still under development.
}
\examples{
# not run:
# dbSingleDF(vSelect = c(var1, var5), filePath = 'exampleDB.db')

}
