% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlite_keywords.R
\docType{data}
\name{sqlite_keywords}
\alias{sqlite_keywords}
\title{Vector of SQLite Keywords.}
\format{
A character vector.
}
\source{
\url{https://www.sqlite.org/c3ref/keyword_check.html}
}
\usage{
sqlite_keywords
}
\description{
A character vector containing the current SQLite Keywords. These strings can not be used as variable names in \code{eatDB}.
}
\keyword{datasets}
