% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_DB_creation.R
\name{createDB}
\alias{createDB}
\title{Create a (relational) data base.}
\usage{
createDB(dfList, pkList, fkList = NULL, metaData = NULL, filePath)
}
\arguments{
\item{dfList}{Named list of data frames. The order of the data frames determines the merge order.}

\item{pkList}{Named list of the primary keys corresponding to the data frames.}

\item{fkList}{Named list of a list per data frame, including referenced data frame ("References") and the corresponding keys ("Keys"). NULL if only a single data frame is supplied.}

\item{metaData}{Data frame including meta data information about the other data frames.}

\item{filePath}{Path to the db file to write (including name); has to end on '.db'.}
}
\value{
Creates a data base in the given path, returns NULL.
}
\description{
Creates a relational data base from a list of data frames. The list structure of dfList, pkList and fkList needs to be exactly the same. Keys (pkList and fkList) can either be single variables or combinations of variables. Primary keys (pkList) have to be unique within a single data table. Foreign Keys (fkList) have to consist of a list with the referenced data frame and the referencing keys. All list elements have to be named and in the same order. If a single data frame is to be converted to a data base, fkList can be dropped.
}
\details{
Primary keys guarantee uniqueness of cases within a single data table, and are single variables or combinations of variables. Foreign keys are used to merge data tables. The foreign key for the first data set is always list(References = NULL, Keys = NULL) The order in which the data frames are supplied determines the merge order. Currently, only Left Joins are implemented. Primary keys have to be variables or combinations of variables that are unique within the data frame.
Data frames are stored seperatly as a relational data base and are merged if pulled from the data base. Then, data frames are joined in the order in which the data frames were supplied in the list. Left Joints are used for merging. SQLite3 is used as a database system.
}
\examples{
# Set up data frames
df1 <- data.frame(v1 = 1, ID2 = 1)
df2 <- data.frame(ID2 = 1:1, v2 = 1:2)
dfList <- list(df1 = df1, df2 = df2)

# Specify primary keys
pkList <- list(df1 = "v1", df2 = c("ID2", "v2"))
# Specify foreign keys
fkList <- list(df1 = list(References = NULL, Keys = NULL),
               df2 = list(References = "df1", Keys = "ID2"))

# Create in memory data base
createDB(dfList = dfList, pkList = pkList, fkList = fkList, metaData = NULL, filePath = ":memory:")

}
